+++ internal/distro/rhel8/distro.go
@@ -929,235 +929,6 @@ func newDistro(distroName string) distro.Distro {
 		basePartitionTables: defaultBasePartitionTables,
 	}
 
-	vhdImgType := imageType{
-		name:     "vhd",
-		filename: "disk.vhd",
-		mimeType: "application/x-vhd",
-		packageSets: map[string]packageSetFunc{
-			buildPkgsKey: distroBuildPackageSet,
-			osPkgsKey:    vhdCommonPackageSet,
-		},
-		packageSetChains: map[string][]string{
-			osPkgsKey: {osPkgsKey, blueprintPkgsKey},
-		},
-		defaultImageConfig: &distro.ImageConfig{
-			EnabledServices: []string{
-				"sshd",
-				"waagent",
-			},
-			DefaultTarget: common.StringToPtr("multi-user.target"),
-		},
-		kernelOptions:       "ro biosdevname=0 rootdelay=300 console=ttyS0 earlyprintk=ttyS0 net.ifnames=0",
-		bootable:            true,
-		defaultSize:         4 * common.GibiByte,
-		pipelines:           vhdPipelines(false),
-		buildPipelines:      []string{"build"},
-		payloadPipelines:    []string{"os", "image", "vpc"},
-		exports:             []string{"vpc"},
-		basePartitionTables: defaultBasePartitionTables,
-	}
-
-	azureRhuiImgType := imageType{
-		name:     "azure-rhui",
-		filename: "disk.vhd.xz",
-		mimeType: "application/xz",
-		packageSets: map[string]packageSetFunc{
-			buildPkgsKey: ec2BuildPackageSet,
-			osPkgsKey:    azureRhuiCommonPackageSet,
-		},
-		packageSetChains: map[string][]string{
-			osPkgsKey: {osPkgsKey, blueprintPkgsKey},
-		},
-		defaultImageConfig: &distro.ImageConfig{
-			Timezone: common.StringToPtr("Etc/UTC"),
-			Locale:   common.StringToPtr("en_US.UTF-8"),
-			GPGKeyFiles: []string{
-				"/etc/pki/rpm-gpg/RPM-GPG-KEY-microsoft-azure-release",
-				"/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-release",
-			},
-			Keyboard: &osbuild.KeymapStageOptions{
-				Keymap: "us",
-				X11Keymap: &osbuild.X11KeymapOptions{
-					Layouts: []string{"us"},
-				},
-			},
-			Sysconfig: []*osbuild.SysconfigStageOptions{
-				{
-					Kernel: &osbuild.SysconfigKernelOptions{
-						UpdateDefault: true,
-						DefaultKernel: "kernel-core",
-					},
-					Network: &osbuild.SysconfigNetworkOptions{
-						Networking: true,
-						NoZeroConf: true,
-					},
-				},
-			},
-			EnabledServices: []string{
-				"firewalld",
-				"nm-cloud-setup.service",
-				"nm-cloud-setup.timer",
-				"sshd",
-				"systemd-resolved",
-				"waagent",
-			},
-			SshdConfig: &osbuild.SshdConfigStageOptions{
-				Config: osbuild.SshdConfigConfig{
-					ClientAliveInterval: common.IntToPtr(180),
-				},
-			},
-			Modprobe: []*osbuild.ModprobeStageOptions{
-				{
-					Filename: "blacklist-amdgpu.conf",
-					Commands: osbuild.ModprobeConfigCmdList{
-						osbuild.NewModprobeConfigCmdBlacklist("amdgpu"),
-					},
-				},
-				{
-					Filename: "blacklist-intel-cstate.conf",
-					Commands: osbuild.ModprobeConfigCmdList{
-						osbuild.NewModprobeConfigCmdBlacklist("intel_cstate"),
-					},
-				},
-				{
-					Filename: "blacklist-floppy.conf",
-					Commands: osbuild.ModprobeConfigCmdList{
-						osbuild.NewModprobeConfigCmdBlacklist("floppy"),
-					},
-				},
-				{
-					Filename: "blacklist-nouveau.conf",
-					Commands: osbuild.ModprobeConfigCmdList{
-						osbuild.NewModprobeConfigCmdBlacklist("nouveau"),
-						osbuild.NewModprobeConfigCmdBlacklist("lbm-nouveau"),
-					},
-				},
-				{
-					Filename: "blacklist-skylake-edac.conf",
-					Commands: osbuild.ModprobeConfigCmdList{
-						osbuild.NewModprobeConfigCmdBlacklist("skx_edac"),
-					},
-				},
-			},
-			CloudInit: []*osbuild.CloudInitStageOptions{
-				{
-					Filename: "10-azure-kvp.cfg",
-					Config: osbuild.CloudInitConfigFile{
-						Reporting: &osbuild.CloudInitConfigReporting{
-							Logging: &osbuild.CloudInitConfigReportingHandlers{
-								Type: "log",
-							},
-							Telemetry: &osbuild.CloudInitConfigReportingHandlers{
-								Type: "hyperv",
-							},
-						},
-					},
-				},
-				{
-					Filename: "91-azure_datasource.cfg",
-					Config: osbuild.CloudInitConfigFile{
-						Datasource: &osbuild.CloudInitConfigDatasource{
-							Azure: &osbuild.CloudInitConfigDatasourceAzure{
-								ApplyNetworkConfig: false,
-							},
-						},
-						DatasourceList: []string{
-							"Azure",
-						},
-					},
-				},
-			},
-			PwQuality: &osbuild.PwqualityConfStageOptions{
-				Config: osbuild.PwqualityConfConfig{
-					Minlen:   common.IntToPtr(6),
-					Minclass: common.IntToPtr(3),
-					Dcredit:  common.IntToPtr(0),
-					Ucredit:  common.IntToPtr(0),
-					Lcredit:  common.IntToPtr(0),
-					Ocredit:  common.IntToPtr(0),
-				},
-			},
-			WAAgentConfig: &osbuild.WAAgentConfStageOptions{
-				Config: osbuild.WAAgentConfig{
-					RDFormat:     common.BoolToPtr(false),
-					RDEnableSwap: common.BoolToPtr(false),
-				},
-			},
-			RHSMConfig: map[distro.RHSMSubscriptionStatus]*osbuild.RHSMStageOptions{
-				distro.RHSMConfigNoSubscription: {
-					DnfPlugins: &osbuild.RHSMStageOptionsDnfPlugins{
-						SubscriptionManager: &osbuild.RHSMStageOptionsDnfPlugin{
-							Enabled: false,
-						},
-					},
-					SubMan: &osbuild.RHSMStageOptionsSubMan{
-						Rhsmcertd: &osbuild.SubManConfigRHSMCERTDSection{
-							AutoRegistration: common.BoolToPtr(true),
-						},
-						Rhsm: &osbuild.SubManConfigRHSMSection{
-							ManageRepos: common.BoolToPtr(false),
-						},
-					},
-				},
-				distro.RHSMConfigWithSubscription: {
-					SubMan: &osbuild.RHSMStageOptionsSubMan{
-						Rhsmcertd: &osbuild.SubManConfigRHSMCERTDSection{
-							AutoRegistration: common.BoolToPtr(true),
-						},
-						// do not disable the redhat.repo management if the user
-						// explicitly request the system to be subscribed
-					},
-				},
-			},
-			Grub2Config: &osbuild.GRUB2Config{
-				TerminalInput:  []string{"serial", "console"},
-				TerminalOutput: []string{"serial", "console"},
-				Serial:         "serial --speed=115200 --unit=0 --word=8 --parity=no --stop=1",
-				Timeout:        10,
-			},
-			UdevRules: &osbuild.UdevRulesStageOptions{
-				Filename: "/etc/udev/rules.d/68-azure-sriov-nm-unmanaged.rules",
-				Rules: osbuild.UdevRules{
-					osbuild.UdevRuleComment{
-						Comment: []string{
-							"Accelerated Networking on Azure exposes a new SRIOV interface to the VM.",
-							"This interface is transparently bonded to the synthetic interface,",
-							"so NetworkManager should just ignore any SRIOV interfaces.",
-						},
-					},
-					osbuild.NewUdevRule(
-						[]osbuild.UdevKV{
-							{K: "SUBSYSTEM", O: "==", V: "net"},
-							{K: "DRIVERS", O: "==", V: "hv_pci"},
-							{K: "ACTION", O: "==", V: "add"},
-							{K: "ENV", A: "NM_UNMANAGED", O: "=", V: "1"},
-						},
-					),
-				},
-			},
-			SystemdUnit: []*osbuild.SystemdUnitStageOptions{
-				{
-					Unit:   "nm-cloud-setup.service",
-					Dropin: "10-rh-enable-for-azure.conf",
-					Config: osbuild.SystemdServiceUnitDropin{
-						Service: &osbuild.SystemdUnitServiceSection{
-							Environment: "NM_CLOUD_SETUP_AZURE=yes",
-						},
-					},
-				},
-			},
-			DefaultTarget: common.StringToPtr("multi-user.target"),
-		},
-		kernelOptions:       "ro crashkernel=auto console=tty1 console=ttyS0 earlyprintk=ttyS0 rootdelay=300",
-		bootable:            true,
-		defaultSize:         68719476736,
-		pipelines:           vhdPipelines(true),
-		buildPipelines:      []string{"build"},
-		payloadPipelines:    []string{"os", "image", "vpc", "archive"},
-		exports:             []string{"archive"},
-		basePartitionTables: azureRhuiBasePartitionTables,
-	}
-
 	vmdkImgType := imageType{
 		name:     "vmdk",
 		filename: "disk.vmdk",
