+++ internal/distro/rhel8/package_sets.go
@@ -376,117 +376,6 @@ func qcow2CommonPackageSet(t *imageType) rpmmd.PackageSet {
 	return ps
 }
 
-func vhdCommonPackageSet(t *imageType) rpmmd.PackageSet {
-	return rpmmd.PackageSet{
-		Include: []string{
-			// Defaults
-			"@Core",
-			"langpacks-en",
-
-			// From the lorax kickstart
-			"chrony",
-			"cloud-init",
-			"cloud-utils-growpart",
-			"gdisk",
-			"net-tools",
-			"python3",
-			"selinux-policy-targeted",
-			"WALinuxAgent",
-
-			// removed from defaults but required to boot in azure
-			"dhcp-client",
-		},
-		Exclude: []string{
-			"dracut-config-rescue",
-			"rng-tools",
-		},
-	}.Append(bootPackageSet(t))
-}
-
-func azureRhuiCommonPackageSet(t *imageType) rpmmd.PackageSet {
-	return rpmmd.PackageSet{
-		Include: []string{
-			"@Server",
-			"NetworkManager",
-			"NetworkManager-cloud-setup",
-			"kernel",
-			"kernel-core",
-			"kernel-modules",
-			"selinux-policy-targeted",
-			"efibootmgr",
-			"lvm2",
-			"grub2-efi-x64",
-			"shim-x64",
-			"dracut-config-generic",
-			"dracut-norescue",
-			"bzip2",
-			"langpacks-en",
-			"grub2-pc",
-			"rhc",
-			"yum-utils",
-			"rhui-azure-rhel8",
-			"WALinuxAgent",
-			"cloud-init",
-			"cloud-utils-growpart",
-			"gdisk",
-			"hyperv-daemons",
-			"nvme-cli",
-			"cryptsetup-reencrypt",
-			"uuid",
-			"rng-tools",
-			"patch",
-		},
-		Exclude: []string{
-			"aic94xx-firmware",
-			"alsa-firmware",
-			"alsa-lib",
-			"alsa-sof-firmware",
-			"alsa-tools-firmware",
-			"dracut-config-rescue",
-			"ivtv-firmware",
-			"iwl1000-firmware",
-			"iwl100-firmware",
-			"iwl105-firmware",
-			"iwl135-firmware",
-			"iwl2000-firmware",
-			"iwl2030-firmware",
-			"iwl3160-firmware",
-			"iwl3945-firmware",
-			"iwl4965-firmware",
-			"iwl5000-firmware",
-			"iwl5150-firmware",
-			"iwl6000-firmware",
-			"iwl6000g2a-firmware",
-			"iwl6000g2b-firmware",
-			"iwl6050-firmware",
-			"iwl7260-firmware",
-			"libertas-sd8686-firmware",
-			"libertas-sd8787-firmware",
-			"libertas-usb8388-firmware",
-			"glibc-all-langpacks",
-			"biosdevname",
-			"cockpit-podman",
-			"bolt",
-			"buildah",
-			"containernetworking-plugins",
-			"dnf-plugin-spacewalk",
-			"iprutils",
-			"plymouth",
-			"podman",
-			"python3-dnf-plugin-spacewalk",
-			"python3-rhnlib",
-			"python3-hwdata",
-			"NetworkManager-config-server",
-			"rhn-client-tools",
-			"rhn-setup",
-			"rhnsd",
-			"rhn-check",
-			"rhnlib",
-			"usb_modeswitch",
-		},
-	}.Append(bootPackageSet(t))
-}
-
 func vmdkCommonPackageSet(t *imageType) rpmmd.PackageSet {
 	return rpmmd.PackageSet{
 		Include: []string{
