+++ internal/distro/rhel8/partition_tables.go
@@ -318,114 +318,3 @@ var edgeBasePartitionTables = distro.BasePartitionTableMap{
 		},
 	},
 }
-
-var azureRhuiBasePartitionTables = distro.BasePartitionTableMap{
-	distro.X86_64ArchName: disk.PartitionTable{
-		UUID: "D209C89E-EA5E-4FBD-B161-B461CCE297E0",
-		Type: "gpt",
-		Size: 68719476736,
-		Partitions: []disk.Partition{
-			{
-				Size: 524288000,
-				Type: disk.EFISystemPartitionGUID,
-				UUID: disk.EFISystemPartitionUUID,
-				Payload: &disk.Filesystem{
-					Type:         "vfat",
-					UUID:         disk.EFIFilesystemUUID,
-					Mountpoint:   "/boot/efi",
-					FSTabOptions: "defaults,uid=0,gid=0,umask=077,shortname=winnt",
-					FSTabFreq:    0,
-					FSTabPassNo:  2,
-				},
-			},
-			{
-				Size: 524288000,
-				Type: disk.FilesystemDataGUID,
-				UUID: disk.FilesystemDataUUID,
-				Payload: &disk.Filesystem{
-					Type:         "xfs",
-					Mountpoint:   "/boot",
-					FSTabOptions: "defaults",
-					FSTabFreq:    0,
-					FSTabPassNo:  0,
-				},
-			},
-			{
-				Size:     2097152,
-				Bootable: true,
-				Type:     disk.BIOSBootPartitionGUID,
-				UUID:     disk.BIOSBootPartitionUUID,
-			},
-			{
-				Type: disk.LVMPartitionGUID,
-				UUID: disk.RootPartitionUUID,
-				Payload: &disk.LVMVolumeGroup{
-					Name:        "rootvg",
-					Description: "built with lvm2 and osbuild",
-					LogicalVolumes: []disk.LVMLogicalVolume{
-						{
-							Size: 1 * 1024 * 1024 * 1024,
-							Name: "homelv",
-							Payload: &disk.Filesystem{
-								Type:         "xfs",
-								Label:        "home",
-								Mountpoint:   "/home",
-								FSTabOptions: "defaults",
-								FSTabFreq:    0,
-								FSTabPassNo:  0,
-							},
-						},
-						{
-							Size: 2 * 1024 * 1024 * 1024,
-							Name: "rootlv",
-							Payload: &disk.Filesystem{
-								Type:         "xfs",
-								Label:        "root",
-								Mountpoint:   "/",
-								FSTabOptions: "defaults",
-								FSTabFreq:    0,
-								FSTabPassNo:  0,
-							},
-						},
-						{
-							Size: 2 * 1024 * 1024 * 1024,
-							Name: "tmplv",
-							Payload: &disk.Filesystem{
-								Type:         "xfs",
-								Label:        "tmp",
-								Mountpoint:   "/tmp",
-								FSTabOptions: "defaults",
-								FSTabFreq:    0,
-								FSTabPassNo:  0,
-							},
-						},
-						{
-							Size: 10 * 1024 * 1024 * 1024,
-							Name: "usrlv",
-							Payload: &disk.Filesystem{
-								Type:         "xfs",
-								Label:        "usr",
-								Mountpoint:   "/usr",
-								FSTabOptions: "defaults",
-								FSTabFreq:    0,
-								FSTabPassNo:  0,
-							},
-						},
-						{
-							Size: 10 * 1024 * 1024 * 1024,
-							Name: "varlv",
-							Payload: &disk.Filesystem{
-								Type:         "xfs",
-								Label:        "var",
-								Mountpoint:   "/var",
-								FSTabOptions: "defaults",
-								FSTabFreq:    0,
-								FSTabPassNo:  0,
-							},
-						},
-					},
-				},
-			},
-		},
-	},
-}
