+++ internal/distro/rhel8/pipelines.go
@@ -54,49 +54,6 @@ func prependKernelCmdlineStage(pipeline *osbuild.Pipeline, t *imageType, pt *dis
 	return pipeline
 }
 
-func vhdPipelines(compress bool) pipelinesFunc {
-	return func(t *imageType, customizations *blueprint.Customizations, options distro.ImageOptions, repos []rpmmd.RepoConfig, packageSetSpecs map[string][]rpmmd.PackageSpec, containers []container.Spec, rng *rand.Rand) ([]osbuild.Pipeline, error) {
-		pipelines := make([]osbuild.Pipeline, 0)
-		pipelines = append(pipelines, *buildPipeline(repos, packageSetSpecs[buildPkgsKey], t.arch.distro.runner))
-
-		partitionTable, err := t.getPartitionTable(customizations.GetFilesystems(), options, rng)
-		if err != nil {
-			return nil, err
-		}
-
-		treePipeline, err := osPipeline(t, repos, packageSetSpecs[osPkgsKey], containers, customizations, options, partitionTable)
-		if err != nil {
-			return nil, err
-		}
-		pipelines = append(pipelines, *treePipeline)
-
-		diskfile := "disk.img"
-		kernelVer := rpmmd.GetVerStrFromPackageSpecListPanic(packageSetSpecs[osPkgsKey], customizations.GetKernel().Name)
-		imagePipeline := liveImagePipeline(treePipeline.Name, diskfile, partitionTable, t.arch, kernelVer)
-		pipelines = append(pipelines, *imagePipeline)
-		if err != nil {
-			return nil, err
-		}
-
-		var qemufile string
-		if compress {
-			qemufile = "disk.vhd"
-		} else {
-			qemufile = t.filename
-		}
-
-		qemuPipeline := qemuPipeline(imagePipeline.Name, diskfile, qemufile, osbuild.QEMUFormatVPC, nil)
-		pipelines = append(pipelines, *qemuPipeline)
-
-		if compress {
-			lastPipeline := pipelines[len(pipelines)-1]
-			pipelines = append(pipelines, *xzArchivePipeline(lastPipeline.Name, qemufile, t.Filename()))
-		}
-
-		return pipelines, nil
-	}
-}
-
 func vmdkPipelines(t *imageType, customizations *blueprint.Customizations, options distro.ImageOptions, repos []rpmmd.RepoConfig, packageSetSpecs map[string][]rpmmd.PackageSpec, containers []container.Spec, rng *rand.Rand) ([]osbuild.Pipeline, error) {
 	pipelines := make([]osbuild.Pipeline, 0)
 	pipelines = append(pipelines, *buildPipeline(repos, packageSetSpecs[buildPkgsKey], t.arch.distro.runner))
